+++ common/sw_canvas.h
@@ -39,7 +39,7 @@ SpiceCanvas *canvas_create(int width, int height, uint32_t format
                            , SpiceZlibDecoder *zlib_decoder
                            );
 
-SpiceCanvas *canvas_create_for_data(int width, int height, uint32_t format, uint8_t *data, size_t stride
+SpiceCanvas *canvas_create_for_data(int width, int height, uint32_t format, uint8_t *data, int stride
 #ifdef SW_CANVAS_CACHE
                            , SpiceImageCache *bits_cache
                            , SpicePaletteCache *palette_cache
